Split caches into separate files modules
