+++ packages/liveblocks-client/src/__tests__/OthersRef.test.ts
@@ -1,4 +1,4 @@
-import { MeRef, OthersPresence } from "../Presence";
+import { OthersRef } from "../OthersRef";
 
 type P = {
   x: number;
@@ -13,37 +13,10 @@ type M = {
   };
 };
 
-describe("Presence", () => {
-  it("empty", () => {
-    const me = new MeRef({ x: 0, y: 0, z: undefined });
-    expect(me.current).toStrictEqual({ x: 0, y: 0 });
-  });
-
-  describe('Tracking "me"', () => {
-    it("patching me", () => {
-      const me = new MeRef<P>({ x: 0, y: 0 });
-      me.patch({ y: 1, z: 2 });
-
-      expect(me.current).toStrictEqual({ x: 0, y: 1, z: 2 });
-    });
-
-    it("patching me with undefineds deletes keys", () => {
-      const me = new MeRef<P>({ x: 1, y: 2 });
-
-      me.patch({ x: undefined });
-      expect(me.current).toStrictEqual({ y: 2 });
-
-      me.patch({ y: undefined });
-      expect(me.current).toStrictEqual({});
-
-      me.patch({ z: undefined });
-      expect(me.current).toStrictEqual({});
-    });
-  });
-
+describe('Read-only "others" ref cache', () => {
   describe('Tracking "others"', () => {
     it("setting alone is not enough other", () => {
-      const p = new OthersPresence<P, M>();
+      const p = new OthersRef<P, M>();
       p.setOther(2, { x: 1, y: 1 });
 
       expect(p.others).toStrictEqual(
@@ -61,7 +34,7 @@ describe("Presence", () => {
     });
 
     it("setting other", () => {
-      const p = new OthersPresence<P, M>();
+      const p = new OthersRef<P, M>();
       p.setConnection(2, "user-123", undefined);
       p.setConnection(3, "user-567", undefined);
 
@@ -76,7 +49,7 @@ describe("Presence", () => {
     });
 
     it("setting others removes explicitly-undefined keys", () => {
-      const p = new OthersPresence<P, M>();
+      const p = new OthersRef<P, M>();
       p.setConnection(2, "user-123", undefined);
       p.setOther(2, { x: 2, y: 2, z: undefined });
       //                             ^^^^^^^^^ 🔑
@@ -88,7 +61,7 @@ describe("Presence", () => {
     });
 
     it("patching others ignores patches for unknown users", () => {
-      const p = new OthersPresence<P, M>();
+      const p = new OthersRef<P, M>();
       p.setConnection(2, "user-123", undefined);
       p.patchOther(2, { y: 1, z: 2 }); // .setOther() not called yet for actor 2
 
@@ -96,7 +69,7 @@ describe("Presence", () => {
     });
 
     it("patching others", () => {
-      const p = new OthersPresence<P, M>();
+      const p = new OthersRef<P, M>();
       p.setConnection(2, "user-123", undefined);
       p.setOther(2, { x: 2, y: 2 });
       expect(p.others).toStrictEqual([
@@ -115,7 +88,7 @@ describe("Presence", () => {
     });
 
     it("removing connections", () => {
-      const p = new OthersPresence<P, M>();
+      const p = new OthersRef<P, M>();
       p.setConnection(2, "user-123", undefined);
       p.setOther(2, { x: 2, y: 2 });
 
@@ -137,35 +110,8 @@ describe("Presence", () => {
   });
 
   describe("caching", () => {
-    it("caches immutable results (me)", () => {
-      const me = new MeRef<P>({ x: 0, y: 0 });
-
-      const me1 = me.current;
-      const me2 = me.current;
-      expect(me1).toBe(me2);
-
-      // These are effectively no-ops
-      me.patch({ x: 0 });
-      me.patch({ y: 0, z: undefined });
-
-      const me3 = me.current;
-      expect(me2).toBe(me3); // No observable change!
-
-      me.patch({ y: -1 });
-
-      const me4 = me.current;
-      const me5 = me.current;
-      expect(me3).not.toBe(me4); // Me changed...
-      expect(me4).toBe(me5);
-
-      const me6 = me.current;
-      const me7 = me.current;
-      expect(me5).toBe(me6); // Me did not change
-      expect(me6).toBe(me7);
-    });
-
     it("caches immutable results (others)", () => {
-      const p = new OthersPresence<P, M>();
+      const p = new OthersRef<P, M>();
       p.setConnection(2, "user-123", undefined);
       p.setOther(2, { x: 2, y: 2 });
 
@@ -194,7 +140,7 @@ describe("Presence", () => {
     });
 
     it("getUser() returns stable cache results", () => {
-      const p = new OthersPresence<P, M>();
+      const p = new OthersRef<P, M>();
       p.setConnection(2, "user-123", undefined);
       p.setOther(2, { x: 2, y: 2 });
